+++ internal/shim/uninstall.go
@@ -5,7 +5,7 @@ import (
 	"log/slog"
 	"os"
 
-	"github.com/spinkube/runtime-class-manager/pkg/state"
+	"github.com/spinkube/runtime-class-manager/internal/state"
 )
 
 func (c *Config) Uninstall(shimName string) (string, error) {
